fix:fabric_gen: Bel port names with numbers get cut off
Bel ports like C0_A and C1_B get cut off during instantiation, since the
rgex that was used to connect the ports as vector, matched any number in
a port name.

This was fixed by replacing the initial regex with one that only matches
trailing numbers in a port name and cutting off the number afterward.

Also, the handling for internal and external ports got combined, since
they are handled the same.

Signed-off-by: Jonas K. <j0nask@proton.me>