Add generative IOs  (#380)
* fix: Custom plib not working with VHDL Makefile

Fix syntax for yosys call with custom plib in Makefile.

Signed-off-by: Jonas K. <j0nask@proton.me>

* feat: Extend Code Generation

Add attributes to addPortVector and addPortSalar, to allow FABulous
attributes to be added to the port description.

Add support for registers.

Add support for inverted ports, wires and registers.

Add support to add BelMaps to a module.

Signed-off-by: Jonas K. <j0nask@proton.me>

* feat: Add generative IOs

Add Gen_IO feature.
Docs in separate commit.

Signed-off-by: Jonas K. <j0nask@proton.me>

* API/CLI: Add support for generative IOs

Signed-off-by: Jonas K. <j0nask@proton.me>

* docs: Add generative IO docs

Signed-off-by: Jonas K. <j0nask@proton.me>

* feat: Add unused tile dict to Fabric class

Unused tiles are listed in the fabric csv, but not used in the Fabric
itself. It can be useful to have these, if one works tile based, and
haven't defined a floorplan yet.

Signed-off-by: Jonas K. <j0nask@proton.me>

* fix:tile_gen: Add bel comment only if there are bels in the tile

Signed-off-by: Jonas K. <j0nask@proton.me>

* chore: Move genIOBel to fabric_automation and make it standalone

Move genIOBel from fabric_gen to fabric_automation.

Add own writer instance to genIOBel, to avoid passing a writer instance.

Adapt changes in API.

Signed-off-by: Jonas K. <j0nask@proton.me>

* feat: Tile: Make globalConfigBits a class method

Make the globalConfigBits a property method,
that calculates the tile config bits when accessed.

Signed-off-by: Jonas K. <j0nask@proton.me>

* docs: Move GEN_IO docs from fabric_definition to fabric_automation

Similar to the code, we move the GEN_IO docs to fabric_automation.

Add some minor details in the example.

Signed-off-by: Jonas K. <j0nask@proton.me>

* fix: build docs workflow

The sphinx-prompt had an update which bricked it, so we fixed the
version to the last one working.

Signed-off-by: Jonas K. <j0nask@proton.me>

---------

Signed-off-by: Jonas K. <j0nask@proton.me>