feat: Improve typer Usage and testing (#475)
Added a lot of testing for Typer front end

Impoving the UX of the interface. 

Instead of calling 
```bash
FABulous run <project> "cmd"
```

Now become if you are not starting in the project dir
```bash
FABulous -p <project> run "cmd"
```

And if is in the project dir becomet;
```bash
FABulous run "cmd"
```

Which is much closer to what typical CLI tool expect and work with. 

------------------------------------------------------------------

* chore: extend argument test

* chore: a few more clean up

* chore: better and more test for CLI

* chore: a few more test case

* chore: partial update

* chore: updated command format

* chore: small update

* feat: update implementation and testing

* chore: Update doc for the new command line

* chore: minor wanning message update

* fix: fix test error

* chore: fix and remove redundant test

* feat: confirmation for overwrite

* chore: fix doc

* chore: remove create project ext test

* chore: update test duration fallback

* fix: fix run command started in interactive mode

* chore: fix doc

* feat: the FABulous <project_dir> for starting

* chore: pre-commit